Stop leaking processes (making Zombies)
We have to call Wait() ourselves as the cmd object going out of scope
does not do that for us. Also, we should make sure stdout of the called
process is closed when leaving scope. Finally, be a bit more aggressive
in returning nothing to the processing pipeline if there is a process
error.